requirements.txt
# SentinelLLM Requirements - Python 3.10 Compatible
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Google Cloud
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0

# Datadog (alternative to OpenTelemetry Datadog exporter)
datadog-api-client==2.16.0

# Security and validation
cryptography==41.0.7
validators==0.22.0

# Logging and monitoring
structlog==23.2.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0

# Environment management
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.10.0
mypy>=1.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# Additional utilities
python-multipart==0.0.6
rich>=13.0.0
click>=8.0.0
typer>=0.7.0

